refactor(display): extract print_track_info into display module
Add rich dependency and move PrintableField, PRINT_WIDTHS, PRINT_HEADERS,
truncate_field, and print_track_info from utils to a new display module
with a module-level Console for upcoming rich-based rendering. Update
edit, convert, and search command imports. Move associated tests from
test_utils.py to test_display.py. Pure move; no behavior change.
