+++ pyproject.toml
@@ -16,6 +16,7 @@ dependencies = [
     "ffmpeg-python>=0.2.0",
     "click>=8.0.0,<=9.0.0",
     "platformdirs>=4.3.8,<5.0.0",
+    "rich>=13.0.0,<15.0.0",
 ]
 
 [dependency-groups]
